repository: URL that can be harvested are assumed to be preferred
When the URL of a source code repository is discovered in the URL of a
source code repository, it is assumed to be preferred. For instance
when the web display of https://sourceforge.net/p/bzflag/code/HEAD/tree/
has svn://svn.code.sf.net/p/bzflag/code/trunk, the web interface is
preferred.

Change-Id: Ifc681ebcd9df10048bc18e6d474dea0746baafd5
Signed-off-by: Loic Dachary <loic@dachary.org>