bot: implement the Bot base class
The Bot base class automatically fetches entities from attributes such
as

   self.P_protocol
   self.Q_git

If initialized with the test=True argument it works against
test.wikidata.org instead of wikidata.org. If an entity does not exist
in test.wikidata.org, a new one is created with the same label. In the
case of a property, it is created with the same datatype as the property
that has the same label in wikidata.

Change-Id: I0a1f0ddf858cbcd319be7fd6d11a012968be623f
Signed-off-by: Loic Dachary <loic@dachary.org>