bot: handle ambiguous names in search_entity
In some cases there is no other way but to hardcode the item id because
the labels are ambiguous and there is no sure way to figure out which
one is relevant in the context of the FLOSS project.

When there are two items and one of them is a disambiguation page, just
ignore it.

Change-Id: I14b895a6a67dcb7da6246c489d10f1b77e1d5777
Signed-off-by: Loic Dachary <loic@dachary.org>