complete rework for simplification
Having subcommands for qa or repository is not a useful separation since
both should be done on each item. Instead have only one set of options
with qa and repositories as plugins in charge of checking a given aspect
of the item.

The items on which the plugins are applied are selected by the main
function instead of being delegated to qa or repository. The plugins
are only responsible for providing SPARQL queries that can be used to
focus on a specific subset of items instead of working on all of them
all the time.

Change-Id: Id99402a1b3e61eb969a5202cae0587a1cd5775f6
Signed-off-by: Loic Dachary <loic@dachary.org>