plugin: replace point-in-time with retrieved
As discussed at
https://www.wikidata.org/wiki/Wikidata:Requests_for_permissions/Bot/FLOSSbot_3
using point in time, although not completely wrong, is a little
strange. Using retrieved in the source / reference instead is more
suitable. Although it is unclear if retrieved relates to the reference
URL (if any) in the sources or the value of the claim itself, it does
not make a significant difference in practice.

Change-Id: I5957cbfca7b2a9607e47b50301d7413f07cae204
Signed-off-by: Loic Dachary <loic@dachary.org>